format changes and automatically terminates brute force if size of art file is too big

+++ src/Art.h
@@ -88,7 +88,7 @@ public:
        * Set the price of the art piece
        * @param price the price of the art piece
        */
-      void setPrice(double);
+      void setPrice(int);
 
       /**
        * Comparators
